Difference between alcohol values and average

function calculates the average alcohol consumption and servings of selected countries, as well as the difference between the alcohol consumption and servings of each country and the average.

Usage

calculate_country_alcohol(name, type)

Arguments

name

Character. Country name. Note: For Russia, using "Russian Federation".

type

Character. The type of alcohol value, one of "beer_servings", "spirit_servings", "wine_servings" and "total_litres_of_pure_alcohol".

Value

A list. A list of length 2. A Character. Mean of alcohol value of selected country.

A data.frame. Difference alcohol value between country in descending order of difference.

country : Country name.

diff : Difference between alcohol value of country and average alcohol value of selected country.

beer_servings/ spirit_servings/ wine_servings/ total_litres_of_pure_alcohol : value of beer_servings, spirit_servings, wine_servings or total_litres_of_pure_alcohol.

Examples

# calculate average wine servings of China, Japan, France, Australia and USA,
# Calculate difference between wine_servings of each country with the average wine_servings.
country <- c("China","Japan","France", "Australia","USA")
calculate_country_alcohol(country, "wine_servings")
#> [[1]]
#> [1] "mean: 138"
#> 
#> [[2]]
#> # A tibble: 5 x 3
#>   country    diff wine_servings
#>   <chr>     <dbl>         <dbl>
#> 1 France      232           370
#> 2 Australia    74           212
#> 3 USA         -54            84
#> 4 Japan      -122            16
#> 5 China      -130             8
#>